Fix test that had accidentally gotten the updates from a commit in a dev branch
The AsymptoticStability velocity analysis function has been modified in a development branch to allow for multiple system analysis. Thus, the velocity analysis result file name was prefixed with the target system name. This changes had accidentally made it onto the test where the name for the analysis was preceeded by 'aeroelastic_'.