fix [structural solvers]  add absolute tolerance as an input option
The absolute tolerance used in xbeam for convergence test is added as a user defined input to the base structural solver and is passed to xbeam with xbopts. This input allows a consistent definition of the absolute tolerance. Further, its default value is increased as it was defined to small to catch very small residuals, which led to non-convergent simulations in the TestGustAssembly unittest.