fix [modalutils] number of modes to be exported as vtk
In the previous version, less modes are extracted for free flying aircraft because number of rigid body DoFs are wrongly substracted from the total number of modes although the first modes (which are the rigid body modes) are already skipped. Thus if 30 modes are to be exported only 10 mode shapes are created.